requirements.txt
Flask==2.3.4
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
